requirements.txt
# ----------------------
# Flask
# ----------------------
flask
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
psycopg2
beautifulsoup4
requests
bs4
PyJWT

# ----------------------
# Production Server
# ----------------------
gunicorn